new-portfolio/backend/requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.4.6
ipython>=8.26.0
django-extensions>=3.2.3

# Testing
pytest>=8.3.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
factory-boy>=3.3.0

# Linting
flake8>=7.1.0
black>=24.8.0
isort>=5.13.0
